% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovingtimer.R
\docType{package}
\name{lovingtimer}
\alias{lovingtimer}
\title{lovingtimer: Create a loving timer in R}
\usage{
lovingtimer(
  title = "From the first day to the future",
  firstday = "2023-01-01 01:00:00",
  bgcolor = "#8dd3c7!important",
  titlefontfamily = "Source Han Serif",
  timerfontfamily = "Source Han Serif !important",
  spanfontfamily = "Helvetica,Arial,sans-serif",
  width = NULL,
  height = NULL
)
}
\arguments{
\item{title}{Title.}

\item{firstday}{The first day of love. For example, "2023-01-01 01:00:00".}

\item{bgcolor}{background-color.}

\item{titlefontfamily}{Title font family.}

\item{timerfontfamily}{Timer font family.}

\item{spanfontfamily}{Font for numbers.}

\item{width}{weight.}

\item{height}{height.}
}
\description{
Create a loving timer in R
}
\section{\code{\link{lovingtimer}}}{
 Create a loving timer in R
}

\examples{
library(lovingtimer)
if (interactive()) {
  lovingtimer()
}
}
